Fix: QA#15 수정
"모임 완료를 눌렀지만 참여자가 새로고침을 안하고 모임 탈퇴?를 누르면 모임에서 나간 상태로 모임 완료상태가됨"

+++ src/main/java/com/hanamja/moa/api/service/group/GroupService.java
@@ -412,9 +412,16 @@ public class GroupService {
                 .orElseThrow(() -> new NotFoundException(HttpStatus.BAD_REQUEST, "해당 그룹을 찾을 수 없습니다."));
 
         Group group = userGroup.getGroup();
-        group.subtractCurrentPeopleNum();
-        userGroupRepository.delete(userGroup);
-        groupRepository.save(group);
+        if(group.getState().equals(State.RECRUITING)){
+            group.subtractCurrentPeopleNum();
+            userGroupRepository.delete(userGroup);
+            groupRepository.save(group);
+        } else {
+            throw UserInputException.builder()
+                    .httpStatus(HttpStatus.BAD_REQUEST)
+                    .message("모임이 이미 모집 완료되었습니다.")
+                    .build();
+        }
 
         return GroupInfoResponseDto.from(group, getHashtagStringList(group));
     }
